requirements_dev.txt
-r requirements.txt

# dev
ipython
werkzeug
django-debug-toolbar

# auto rebuild/testing
watchdog
git+git://github.com/yejianye/watchdog-tricks.git
